Switched the toggle button B to specific buttons x and y
Changed the toggle button B that switched from Field Centric to Robot Centric to buttons x and y. If x is pressed, then it's switched to Field Centric. If y is pressed, then it switches to Robot Centric. I got rid of the old code that had the toggle button B functions